전투 시스템 확장 및 코드 품질 개선
`CombatCalculator.cs`에서 `CalcDamage`와 `CalcHealing` 메서드의 계수 타입이 `int`에서 `float`로 변경되어 계산 정밀도가 향상되었습니다.
`DamageApplication` 구조체의 `ShieldAborbed` 오타가 `ShieldAbsorbed`로 수정되었습니다.
`ActionResolver.cs` 파일이 추가되어 전투 행동 실행 및 조율을 담당하며, 피해 계산, 회복, 보호막 적용 등의 기능을 제공합니다.

+++ Assets/Scripts/Systems/CombatCalculator.cs
@@ -9,7 +9,7 @@ public static class CombatCalculator
     // 피해 사슬: 기본피해 * 방어보정 * 방어계수 * 붕괴 -> 소수점 버림, 최소 1
     // directionMod, breakMod = 밖에서 판정한 배율. 보정 없으면 1.0f 주입
     public static DamageResult CalcDamage(
-        int attack, int damageCoeffi, int fixedDamage,
+        int attack, float damageCoeffi, int fixedDamage,
         int defense, float directionMod, float breakMod)
     {
         double basePower = attack * (double)damageCoeffi + fixedDamage;         // 공격력, 계수, 고정피해
@@ -21,7 +21,7 @@ public static class CombatCalculator
     }
 
     // 회복: 방어, 방향, 붕괴 무보정. 받는회복량 보정만 밖에서 주임
-    public static int CalcHealing(int attack, int healingCoeffi, int fixedHealing, float receivedHealingMod)
+    public static int CalcHealing(int attack, float healingCoeffi, int fixedHealing, float receivedHealingMod)
     { 
         double raw = (attack * (double)healingCoeffi + fixedHealing) * receivedHealingMod;
         return Math.Max(0, FloorToInt(raw));
